Bugfix: Use pixel center as reference
When computing forces for each screen pixel, use the center of
each pixel as reference coordinates, e.g. (0.5, 0.5) instead
of (0.0, 0.0) for the pixel in the upper left corner.
This change fixes minor rounding errors.  In particular, it
avoids the strange behavior of round shaped walls to continously
track the ball along the rounded shape if the ball is approaching
such a wall in a extremely flat angle.