Bugfix: Consider playing field geometry for ball speed
Internally, the playing field covers the double valued
range [0.0 .. 1.0] for x and y coordinates, i.e. a
quadratic field.  When scaled to a non-quadratic
window or screen size, the ball effectively moves
more quickly in x or y direction due to the different
numbers of pixels in x and y direction.
This commit introduces an x and y direction velocity
correction factor for compensating the effect of
scaling the playing field to a non-quadratic square.